extract pure status math into server/status
